PR-5 Scaffold refactor
